* commit '3de2377d7abb6afcca8acb9a927a25dc99056141':
  sqlite integer fields are long

+++ cmds/content/src/com/android/commands/content/Content.java
@@ -498,7 +498,7 @@ public class Content {
                                     columnValue = String.valueOf(cursor.getFloat(columnIndex));
                                     break;
                                 case Cursor.FIELD_TYPE_INTEGER:
-                                    columnValue = String.valueOf(cursor.getInt(columnIndex));
+                                    columnValue = String.valueOf(cursor.getLong(columnIndex));
                                     break;
                                 case Cursor.FIELD_TYPE_STRING:
                                     columnValue = cursor.getString(columnIndex);
